+++ tools/src/librlink/RlinkChannel.cpp
@@ -1,6 +1,6 @@
-// $Id: RlinkChannel.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RlinkChannel.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -68,7 +68,7 @@ bool RlinkChannel::Exec(RlinkCommandList& clist, RerrMsg& emsg)
 //! FIXME_docs
 
 void RlinkChannel::Dump(std::ostream& os, int ind, const char* text,
-                        int detail) const
+                        int /*detail*/) const
 {
   RosFill bl(ind);
   os << bl << (text?text:"--") << "RlinkChannel @ " << this << endl;