+++ tools/src/librlink/RlinkCommandExpect.cpp
@@ -1,6 +1,6 @@
-// $Id: RlinkCommandExpect.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RlinkCommandExpect.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2011-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2011-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -134,7 +134,7 @@ bool RlinkCommandExpect::BlockIsChecked(size_t ind) const
 //! FIXME_docs
 
 void RlinkCommandExpect::Dump(std::ostream& os, int ind, const char* text,
-                              int detail) const
+                              int /*detail*/) const
 {
   RosFill bl(ind);
   os << bl << (text?text:"--") << "RlinkCommandExpect @ " << this << endl;