+++ tools/src/librutiltpp/RtclSystem.cpp
@@ -1,6 +1,6 @@
-// $Id: RtclSystem.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RtclSystem.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2014 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -69,7 +69,7 @@ void RtclSystem::CreateCmds(Tcl_Interp* interp)
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-int RtclSystem::Isatty(ClientData cdata, Tcl_Interp* interp, 
+int RtclSystem::Isatty(ClientData /*cdata*/, Tcl_Interp* interp, 
                        int objc, Tcl_Obj* const objv[])
 {
   RtclArgs args(interp, objc, objv);
@@ -116,7 +116,7 @@ static const char* signum2nam(int signum)
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-int RtclSystem::SignalAction(ClientData cdata, Tcl_Interp* interp, 
+int RtclSystem::SignalAction(ClientData /*cdata*/, Tcl_Interp* interp, 
                              int objc, Tcl_Obj* const objv[])
 {
   RtclArgs args(interp, objc, objv);
@@ -210,7 +210,7 @@ int RtclSystem::SignalAction(ClientData cdata, Tcl_Interp* interp,
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-int RtclSystem::WaitPid(ClientData cdata, Tcl_Interp* interp, 
+int RtclSystem::WaitPid(ClientData /*cdata*/, Tcl_Interp* interp, 
                         int objc, Tcl_Obj* const objv[])
 {
   RtclArgs args(interp, objc, objv);