+++ tools/src/librw11/Rw11Cntl.cpp
@@ -1,6 +1,6 @@
-// $Id: Rw11Cntl.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11Cntl.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -98,7 +98,7 @@ void Rw11Cntl::Start()
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-bool Rw11Cntl::BootCode(size_t unit, std::vector<uint16_t>& code, 
+  bool Rw11Cntl::BootCode(size_t /*unit*/, std::vector<uint16_t>& code, 
                         uint16_t& aload, uint16_t& astart)
 {
   code.clear();