+++ tools/src/librw11/Rw11CntlDL11.cpp
@@ -1,6 +1,6 @@
-// $Id: Rw11CntlDL11.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11CntlDL11.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -132,7 +132,7 @@ void Rw11CntlDL11::Start()
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-void Rw11CntlDL11::UnitSetup(size_t ind)
+void Rw11CntlDL11::UnitSetup(size_t /*ind*/)
 {
   Rw11Cpu& cpu  = Cpu();
   uint16_t rcsr = (fRxRlim<<kRCSR_V_RXRLIM) & kRCSR_M_RXRLIM;