+++ tools/src/librw11/Rw11CntlPC11.cpp
@@ -1,6 +1,6 @@
-// $Id: Rw11CntlPC11.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11CntlPC11.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -132,7 +132,7 @@ void Rw11CntlPC11::Start()
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-bool Rw11CntlPC11::BootCode(size_t unit, std::vector<uint16_t>& code, 
+bool Rw11CntlPC11::BootCode(size_t /*unit*/, std::vector<uint16_t>& code, 
                             uint16_t& aload, uint16_t& astart)
 {
   uint16_t kBOOT_START = 0017476;