+++ tools/src/librw11/Rw11Unit.cpp
@@ -1,6 +1,6 @@
-// $Id: Rw11Unit.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11Unit.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -89,7 +89,7 @@ const std::string&  Rw11Unit::AttachUrl() const
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-bool Rw11Unit::Attach(const std::string& url, RerrMsg& emsg)
+bool Rw11Unit::Attach(const std::string& /*url*/, RerrMsg& emsg)
 {
   emsg.Init("Rw11Unit::Attach","attach not available for this device type");
   return false;