+++ tools/src/librw11/Rw11VirtTermTcp.cpp
@@ -1,6 +1,6 @@
-// $Id: Rw11VirtTermTcp.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11VirtTermTcp.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -188,7 +188,7 @@ bool Rw11VirtTermTcp::Open(const std::string& url, RerrMsg& emsg)
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-bool Rw11VirtTermTcp::Snd(const uint8_t* data, size_t count, RerrMsg& emsg)
+bool Rw11VirtTermTcp::Snd(const uint8_t* data, size_t count, RerrMsg& /*emsg*/)
 {
   fStats.Inc(kStatNVTSnd);
   const uint8_t* pdata = data;