+++ tools/src/testtclsh/testtclsh.cpp
@@ -1,6 +1,6 @@
-// $Id: testtclsh.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: testtclsh.cpp 1049 2018-09-22 13:56:52Z mueller $
 //
-// Copyright 2013- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -32,7 +32,7 @@ extern "C" int Rlinktpp_Init(Tcl_Interp* interp);
 //extern "C" int Rusbtpp_Init(Tcl_Interp* interp);
 extern "C" int Rwxxtpp_Init(Tcl_Interp* interp);
 
-int main(int argc, const char* argv[]) 
+int main(int /*argc*/, const char* /*argv*/[]) 
 {
   cout << "testtclsh starting..." << endl;